Completed planet property update
Updating all planet related fields on planet info view
Updates affect game play

+++ Assets/Scripts/Data/PlanetData.cs
@@ -9,7 +9,7 @@ public class PlanetData
 	public int ShipCargoLevel;
 	public float CurrentTotalMiningRate;
 	public float CurrentShipSpeed;
-	public float CurrentShipCargo;
+	public int CurrentShipCargo;
 	public int TotalMiningRateUpdatePrice;
 	public int ShipSpeedUpdatePrice;
 	public int ShipCargoUpdatePrice;
